Lots of refactoring... lost count
