+++ android-emojis/emoji-keyboard/src/main/AndroidManifest.xml
@@ -1,8 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.klinker.android.emoji_keyboard_trial"
-          android:versionCode="9"
-          android:versionName="1.37" >
+          package="com.klinker.android.emoji_keyboard_trial" >
 
     <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.VIBRATE"/>
