+++ android-emojis/emoji-keyboard/src/main/java/com/klinker/android/emoji_keyboard/Utility.java
@@ -1,0 +1,14 @@
+package com.klinker.android.emoji_keyboard;
+
+import java.util.ArrayList;
+
+public class Utility {
+    public static final ArrayList<Integer> initArrayList(int... ints) {
+        ArrayList<Integer> list = new ArrayList<Integer>();
+        for (int i : ints)
+        {
+            list.add(i);
+        }
+        return list;
+    }
+}
