+++ android-emojis/emoji-keyboard/src/main/java/com/klinker/android/emoji_keyboard/constants/EmojiTexts.java
@@ -1,0 +1,196 @@
+package com.klinker.android.emoji_keyboard.constants;
+
+public interface EmojiTexts {
+
+    String[] transEmojiTexts = {
+            // transportation
+            "\uD83C\uDFE0", "\uD83C\uDFe1", "\uD83C\uDFE2", "\uD83C\uDFE3", "\uD83C\uDFE4",
+            "\uD83C\uDFE5", "\uD83C\uDFe6", "\uD83C\uDFE7", "\uD83C\uDFE8", "\uD83C\uDFE9",
+            "\uD83C\uDFEa", "\uD83C\uDFeb", "\u26ea", "\u26f2", "\uD83C\uDFEc",
+            "\uD83C\uDFEf", "\uD83C\uDFf0", "\uD83C\uDFEd", "\uD83D\uDDFB", "\uD83D\uDDFc",
+            "\uD83D\uDC88", "\uD83D\uDD27", "\uD83D\uDD28", "\uD83D\uDD29", "\uD83D\uDEBF",
+            "\uD83D\uDEc1", "\uD83D\uDEc0", "\uD83D\uDEbd", "\uD83D\uDEBe", "\uD83C\uDFbd",
+            "\uD83C\uDFa3", "\uD83C\uDFb1", "\uD83C\uDFb3", "\u26be", "\u26f3",
+            "\uD83C\uDFbe", "\u26bd", "\uD83C\uDFbf", "\uD83C\uDFc0", "\uD83C\uDFc1",
+            "\uD83C\uDFc2", "\uD83C\uDFc3", "\uD83C\uDFc4", "\uD83C\uDFc6", "\uD83C\uDFc7",
+            "\uD83D\uDC0E", "\uD83C\uDFc8", "\uD83C\uDFc9", "\uD83C\uDFca", "\uD83D\uDE82",
+            "\uD83D\uDE83", "\uD83D\uDE84", "\uD83D\uDE85", "\uD83D\uDE86", "\uD83D\uDE87",
+            "\u24c2", "\uD83D\uDE88", "\uD83D\uDE8a", "\uD83D\uDE8b", "\uD83D\uDE8c",
+            "\uD83D\uDE8d", "\uD83D\uDE8e", "\uD83D\uDE8f", "\uD83D\uDE90", "\uD83D\uDE91",
+            "\uD83D\uDE92", "\uD83D\uDE93", "\uD83D\uDE94", "\uD83D\uDE95", "\uD83D\uDE96",
+            "\uD83D\uDE97", "\uD83D\uDE98", "\uD83D\uDE99", "\uD83D\uDE9a", "\uD83D\uDE9b",
+            "\uD83D\uDE9c", "\uD83D\uDE9d", "\uD83D\uDE9e", "\uD83D\uDE9f", "\uD83D\uDEa0",
+            "\uD83D\uDEa1", "\uD83D\uDEa2", "\uD83D\uDEa3", "\uD83D\uDE81", "\u2708",
+            "\uD83D\uDEc2", "\uD83D\uDEc3", "\uD83D\uDEc4", "\uD83D\uDEc5", "\u26f5",
+            "\uD83D\uDEb2", "\uD83D\uDEb3", "\uD83D\uDEb4", "\uD83D\uDEb5", "\uD83D\uDEb7",
+            "\uD83D\uDEb8", "\uD83D\uDE89", "\uD83D\uDE80", "\uD83D\uDEa4", "\uD83D\uDEb6",
+            "\u26fd", "\uD83C\uDD7F", "\uD83D\uDEa5", "\uD83D\uDEa6", "\uD83D\uDEa7",
+            "\uD83D\uDEa8", "\u2668", "\uD83D\uDC8C", "\uD83D\uDC8d", "\uD83D\uDC8e",
+            "\uD83D\uDC90", "\uD83D\uDC92", "\uD83C\uDDEF\uD83C\uDDF5", "\uD83C\uDDFA\uD83C\uDDF8", "\uD83C\uDDEB\uD83C\uDDF7",
+            "\uD83C\uDDE9\uD83C\uDDEA", "\uD83C\uDDEE\uD83C\uDDF9", "\uD83C\uDDEC\uD83C\uDDE7", "\uD83C\uDDEA\uD83C\uDDF8", "\uD83C\uDDF7\uD83C\uDDFA",
+            "\uD83C\uDDE8\uD83C\uDDF3", "\uD83C\uDDF0\uD83C\uDDF7"
+    };
+
+    String[] thingsEmojiTexts = {
+            //things
+            "\uD83D\uDD30", "\uD83D\uDC84", "\uD83D\uDC5e", "\uD83D\uDC5f", "\uD83D\uDC51",
+            "\uD83D\uDC52", "\uD83C\uDFa9", "\uD83C\uDF93", "\uD83D\uDC53", "\u231a",
+            "\uD83D\uDC54", "\uD83D\uDC55", "\uD83D\uDC56", "\uD83D\uDC57", "\uD83D\uDC58",
+            "\uD83D\uDC59", "\uD83D\uDC60", "\uD83D\uDC61", "\uD83D\uDC62", "\uD83D\uDC5a",
+            "\uD83D\uDC5c", "\uD83D\uDCbc", "\uD83C\uDF92", "\uD83D\uDC5d", "\uD83D\uDC5b",
+            "\uD83D\uDCb0", "\uD83D\uDCb3", "\uD83D\uDCb2", "\uD83D\uDCb5", "\uD83D\uDCb4",
+            "\uD83D\uDCb6", "\uD83D\uDCb7", "\uD83D\uDCb1", "\uD83D\uDCb8", "\uD83D\uDCb9",
+            "\uD83D\uDD2b", "\uD83D\uDD2a", "\uD83D\uDCa3", "\uD83D\uDC89", "\uD83D\uDC8a",
+            "\uD83D\uDEac", "\uD83D\uDD14", "\uD83D\uDD15", "\uD83D\uDEaa", "\uD83D\uDD2c",
+            "\uD83D\uDD2d", "\uD83D\uDD2e", "\uD83D\uDD26", "\uD83D\uDD0b", "\uD83D\uDD0c",
+            "\uD83D\uDCdc", "\uD83D\uDCd7", "\uD83D\uDCd8", "\uD83D\uDCd9", "\uD83D\uDCda",
+            "\uD83D\uDCd4", "\uD83D\uDCd2", "\uD83D\uDCd1", "\uD83D\uDCd3", "\uD83D\uDCd5",
+            "\uD83D\uDCd6", "\uD83D\uDCf0", "\uD83D\uDCdb", "\uD83C\uDF83", "\uD83C\uDF84",
+            "\uD83C\uDF80", "\uD83C\uDF81", "\uD83C\uDF82", "\uD83C\uDF88", "\uD83C\uDF86",
+            "\uD83C\uDF87", "\uD83C\uDF89", "\uD83C\uDF8a", "\uD83C\uDF8d", "\uD83C\uDF8f",
+            "\uD83C\uDF8c", "\uD83C\uDF90", "\uD83C\uDF8b", "\uD83C\uDF8e", "\uD83D\uDCf1",
+            "\uD83D\uDCf2", "\uD83D\uDCdf", "\u260e", "\uD83D\uDCde", "\uD83D\uDCe0",
+            "\uD83D\uDCe6", "\u2709", "\uD83D\uDCe8", "\uD83D\uDCe9", "\uD83D\uDCea",
+            "\uD83D\uDCeb", "\uD83D\uDCed", "\uD83D\uDCec", "\uD83D\uDCee", "\uD83D\uDCe4",
+            "\uD83D\uDCe5", "\uD83D\uDCef", "\uD83D\uDCe3", "\uD83D\uDCe2", "\uD83D\uDCe1",
+            "\uD83D\uDCac", "\uD83D\uDCad", "\u2712", "\u270f", "\uD83D\uDCdd",
+            "\uD83D\uDCcf", "\uD83D\uDCd0", "\uD83D\uDCcd", "\uD83D\uDCcc", "\uD83D\uDCce",
+            "\u2702", "\uD83D\uDCba", "\uD83D\uDCbb", "\uD83D\uDCbd", "\uD83D\uDCbe",
+            "\uD83D\uDCbf", "\uD83D\uDCc6", "\uD83D\uDCc5", "\uD83D\uDCc7", "\uD83D\uDCcb",
+            "\uD83D\uDCc1", "\uD83D\uDCc2", "\uD83D\uDCc3", "\uD83D\uDCc4", "\uD83D\uDCca",
+            "\uD83D\uDCc8", "\uD83D\uDCc9", "\u26fa", "\uD83C\uDFa1", "\uD83C\uDFa1",
+            "\uD83C\uDFa0", "\uD83C\uDFaa", "\uD83C\uDFa8", "\uD83C\uDFac", "\uD83C\uDFa5",
+            "\uD83D\uDCf7", "\uD83D\uDCf9", "\uD83C\uDFa6", "\uD83C\uDFad", "\uD83C\uDFab",
+            "\uD83C\uDFae", "\uD83C\uDFb2", "\uD83C\uDFb0", "\uD83C\uDCCF", "\uD83C\uDFb4",
+            "\uD83C\uDC04", "\uD83C\uDFaf", "\uD83D\uDCfa", "\uD83D\uDCfb", "\uD83D\uDCc0",
+            "\uD83D\uDCfc", "\uD83C\uDFa7", "\uD83C\uDFa4", "\uD83C\uDFb5", "\uD83C\uDFb6",
+            "\uD83C\uDFbc", "\uD83C\uDFbb", "\uD83C\uDFb9", "\uD83C\uDFb7", "\uD83C\uDFba",
+            "\uD83C\uDFb8", "\u303d"
+    };
+
+    String[] peopleEmojiTexts = {
+            // People
+            "\u263A", "\uD83D\uDE0A", "\uD83D\uDE00", "\uD83D\uDE01", "\uD83D\uDE02",
+            "\uD83D\uDE03", "\uD83D\uDE04", "\uD83D\uDE05", "\uD83D\uDE06", "\uD83D\uDE07",
+            "\uD83D\uDE08", "\uD83D\uDE09", "\uD83D\uDE2F", "\uD83D\uDE10", "\uD83D\uDE11",
+            "\uD83D\uDE15", "\uD83D\uDE20", "\uD83D\uDE2C", "\uD83D\uDE21", "\uD83D\uDE22",
+            "\uD83D\uDE34", "\uD83D\uDE2E", "\uD83D\uDE23", "\uD83D\uDE24", "\uD83D\uDE25",
+            "\uD83D\uDE26", "\uD83D\uDE27", "\uD83D\uDE28", "\uD83D\uDE29", "\uD83D\uDE30",
+            "\uD83D\uDE1F", "\uD83D\uDE31", "\uD83D\uDE32", "\uD83D\uDE33", "\uD83D\uDE35",
+            "\uD83D\uDE36", "\uD83D\uDE37", "\uD83D\uDE1E", "\uD83D\uDE12", "\uD83D\uDE0D",
+            "\uD83D\uDE1b", "\uD83D\uDE1c", "\uD83D\uDE1d", "\uD83D\uDE0b", "\uD83D\uDE17",
+            "\uD83D\uDE19", "\uD83D\uDE18", "\uD83D\uDE1a", "\uD83D\uDE0e", "\uD83D\uDE2d",
+            "\uD83D\uDE0c", "\uD83D\uDE16", "\uD83D\uDE14", "\uD83D\uDE2a", "\uD83D\uDE0f",
+            "\uD83D\uDE13", "\uD83D\uDE2b", "\uD83D\uDE4b", "\uD83D\uDE4c", "\uD83D\uDE4d",
+            "\uD83D\uDE45", "\uD83D\uDE46", "\uD83D\uDE47", "\uD83D\uDE4e", "\uD83D\uDE4f",
+            "\uD83D\uDE3a", "\uD83D\uDE3c", "\uD83D\uDE38", "\uD83D\uDE39", "\uD83D\uDE3b",
+            "\uD83D\uDE3d", "\uD83D\uDE3f", "\uD83D\uDE3e", "\uD83D\uDE40", "\uD83D\uDE48",
+            "\uD83D\uDE49", "\uD83D\uDE4a", "\uD83D\uDCA9", "\uD83D\uDC76", "\uD83D\uDC66",
+            "\uD83D\uDC66", "\uD83D\uDC68", "\uD83D\uDC69", "\uD83D\uDC74", "\uD83D\uDC75",
+            "\uD83D\uDC8f", "\uD83D\uDC91", "\uD83D\uDC6a", "\uD83D\uDC6b", "\uD83D\uDC6c",
+            "\uD83D\uDC6d", "\uD83D\uDC64", "\uD83D\uDC65", "\uD83D\uDC6e", "\uD83D\uDC77",
+            "\uD83D\uDC81", "\uD83D\uDC82", "\uD83D\uDC6f", "\uD83D\uDC70", "\uD83D\uDC78",
+            "\uD83C\uDF85", "\uD83D\uDC7c", "\uD83D\uDC71", "\uD83D\uDC72", "\uD83D\uDC73",
+            "\uD83D\uDC83", "\uD83D\uDC86", "\uD83D\uDC87", "\uD83D\uDC85", "\uD83D\uDC7b",
+            "\uD83D\uDC79", "\uD83D\uDC7a", "\uD83D\uDC7d", "\uD83D\uDC7e", "\uD83D\uDC7f",
+            "\uD83D\uDC80", "\uD83D\uDCaa", "\uD83D\uDC40", "\uD83D\uDC42", "\uD83D\uDC43",
+            "\uD83D\uDC63", "\uD83D\uDC44", "\uD83D\uDC45", "\uD83D\uDC8b", "\u2764",
+            "\uD83D\uDC99", "\uD83D\uDC9a", "\uD83D\uDC9b", "\uD83D\uDC9c", "\uD83D\uDC93",
+            "\uD83D\uDC94", "\uD83D\uDC95", "\uD83D\uDC96", "\uD83D\uDC97", "\uD83D\uDC98",
+            "\uD83D\uDC9d", "\uD83D\uDC9e", "\uD83D\uDC9f", "\uD83D\uDC4d", "\uD83D\uDC4e",
+            "\uD83D\uDC4c", "\u270a", "\u270c", "\u270b", "\uD83D\uDC4a",
+            "\u261d", "\uD83D\uDC46", "\uD83D\uDC47", "\uD83D\uDC48", "\uD83D\uDC49",
+            "\uD83D\uDC4b", "\uD83D\uDC4f", "\uD83D\uDC50"
+    };
+
+    String[] otherEmojiTexts = {
+            // other
+            "\uD83D\uDD1d", "\uD83D\uDD19", "\uD83D\uDD1b", "\uD83D\uDD1c", "\uD83D\uDD1a",
+            "\u23f3", "\u231b", "\u23f0", "\u2648", "\u2649",
+            "\u264a", "\u264b", "\u264c", "\u264d", "\u264e",
+            "\u264f", "\u2650", "\u2651", "\u2652", "\u2653",
+            "\u26ce", "\uD83D\uDD31", "\uD83D\uDD2f", "\uD83D\uDEbb", "\uD83D\uDEae",
+            "\uD83D\uDEaf", "\uD83D\uDEb0", "\uD83D\uDEb1", "\uD83C\uDD70", "\uD83C\uDD71",
+            "\uD83C\uDD8e", "\uD83C\uDD7e", "\uD83D\uDCae", "\uD83D\uDCaf", "\uD83D\uDD20",
+            "\uD83D\uDD21", "\uD83D\uDD22", "\uD83D\uDD23", "\uD83D\uDD24", "\u27bf",
+            "\uD83D\uDCf6", "\uD83D\uDCf3", "\uD83D\uDCf4", "\uD83D\uDCf5", "\uD83D\uDEb9",
+            "\uD83D\uDEba", "\uD83D\uDEbc", "\u267f", "\u267b", "\uD83D\uDEad",
+            "\uD83D\uDEa9", "\u26a0", "\uD83C\uDE01", "\uD83D\uDD1e", "\u26d4",
+            "\uD83C\uDD92", "\uD83C\uDD97", "\uD83C\uDD95", "\uD83C\uDD98", "\uD83C\uDD99",
+            "\uD83C\uDD93", "\uD83C\uDD96", "\uD83C\uDD9a", "\uD83C\uDE32", "\uD83C\uDE33",
+            "\uD83C\uDE34", "\uD83C\uDE35", "\uD83C\uDE36", "\uD83C\uDE37", "\uD83C\uDE38",
+            "\uD83C\uDE39", "\uD83C\uDE02", "\uD83C\uDE3a", "\uD83C\uDE50", "\uD83C\uDE51",
+            "\u3299", "\u00ae", "\u00a9", "\u2122", "\uD83C\uDE1a",
+            "\uD83C\uDE2f", "\u3297", "\u2b55", "\u274c", "\u274e",
+            "\u2139", "\uD83D\uDEab", "\u2705", "\u2714", "\uD83D\uDD17",
+            "\u2734", "\u2733", "\u2795", "\u2796", "\u2716",
+            "\u2797", "\uD83D\uDCa0", "\uD83D\uDCa1", "\uD83D\uDCa4", "\uD83D\uDCa2",
+            "\uD83D\uDD25", "\uD83D\uDCa5", "\uD83D\uDCa6", "\uD83D\uDCa8", "\uD83D\uDCab",
+            "\uD83D\uDD5b", "\uD83D\uDD67", "\uD83D\uDD50", "\uD83D\uDD5c", "\uD83D\uDD51",
+            "\uD83D\uDD5d", "\uD83D\uDD52", "\uD83D\uDD5e", "\uD83D\uDD53", "\uD83D\uDD5f",
+            "\uD83D\uDD54", "\uD83D\uDD60", "\uD83D\uDD54", "\uD83D\uDD61", "\uD83D\uDD56",
+            "\uD83D\uDD62", "\uD83D\uDD57", "\uD83D\uDD63", "\uD83D\uDD58", "\uD83D\uDD64",
+            "\uD83D\uDD59", "\uD83D\uDD65", "\uD83D\uDD5a", "\uD83D\uDD66", "\u2195",
+            "\u2b06", "\u2197", "\u27a1", "\u2198", "\u2b07",
+            "\u2199", "\u2b05", "\u2196", "\u2194", "\u2934",
+            "\u2935", "\u23ea", "\u23e9", "\u23eb", "\u23ec",
+            "\u25c0", "\u25b6", "\uD83D\uDD3d", "\uD83D\uDD3c", "\u2747",
+            "\u2728", "\uD83D\uDD34", "\uD83D\uDD35", "\u26aa", "\u26ab",
+            "\uD83D\uDD33", "\uD83D\uDD32", "\u2b50", "\uD83C\uDF1f", "\uD83C\uDF20",
+            "\u25ab", "\u25aa", "\u25fd", "\u25fe", "\u25fc",
+            "\u25fb", "\u2b1b", "\u2b1c", "\uD83D\uDD39", "\uD83D\uDD38",
+            "\uD83D\uDD37", "\uD83D\uDD36", "\uD83D\uDD3a", "\uD83D\uDD3b", "\uD83D\uDD1f",
+            "\u20e3", "\u2754", "\u2753", "\u2755", "\u2757",
+            "\u203c", "\u2049", "\u3030", "\u27b0", "\u2660",
+            "\u2665", "\u2663", "\u2666", "\uD83C\uDD94", "\uD83D\uDD11",
+            "\u21a9", "\uD83C\uDD91", "\uD83D\uDD0d", "\uD83D\uDD12", "\uD83D\uDD13",
+            "\u21aa", "\uD83D\uDD10", "\u2611", "\uD83D\uDD18", "\uD83D\uDD0e",
+            "\uD83D\uDD16", "\uD83D\uDD0f", "\uD83D\uDD03", "\uD83D\uDD00", "\uD83D\uDD01",
+            "\uD83D\uDD02", "\uD83D\uDD04", "\uD83D\uDCe7", "\uD83D\uDD05", "\uD83D\uDD06",
+            "\uD83D\uDD07", "\uD83D\uDD08", "\uD83D\uDD09", "\uD83D\uDD0a", "\uD83C\uDDe6",
+            "\uD83C\uDDe7", "\uD83C\uDDe8", "\uD83C\uDDe9", "\uD83C\uDDea", "\uD83C\uDDeb",
+            "\uD83C\uDDec", "\uD83C\uDDed", "\uD83C\uDDee", "\uD83C\uDDef", "\uD83C\uDDf0",
+            "\uD83C\uDDf1", "\uD83C\uDDf2", "\uD83C\uDDf3", "\uD83C\uDDf4", "\uD83C\uDDf5",
+            "\uD83C\uDDf6", "\uD83C\uDDf7", "\uD83C\uDDf8", "\uD83C\uDDf9", "\uD83C\uDDfa",
+            "\uD83C\uDDfb", "\uD83C\uDDfc", "\uD83C\uDDfd", "\uD83C\uDDfe", "\uD83C\uDDff"
+    };
+    String[] natureEmojiTexts = {
+            // nature
+            "\uD83D\uDC15", "\uD83D\uDC36", "\uD83D\uDC29", "\uD83D\uDC08", "\uD83D\uDC31",
+            "\uD83D\uDC00", "\uD83D\uDC01", "\uD83D\uDC2d", "\uD83D\uDC39", "\uD83D\uDC22",
+            "\uD83D\uDC07", "\uD83D\uDC30", "\uD83D\uDC13", "\uD83D\uDC14", "\uD83D\uDC23",
+            "\uD83D\uDC24", "\uD83D\uDC25", "\uD83D\uDC26", "\uD83D\uDC0f", "\uD83D\uDC11",
+            "\uD83D\uDC10", "\uD83D\uDC3a", "\uD83D\uDC03", "\uD83D\uDC02", "\uD83D\uDC04",
+            "\uD83D\uDC2e", "\uD83D\uDC34", "\uD83D\uDC17", "\uD83D\uDC16", "\uD83D\uDC37",
+            "\uD83D\uDC3d", "\uD83D\uDC38", "\uD83D\uDC0d", "\uD83D\uDC3c", "\uD83D\uDC27",
+            "\uD83D\uDC18", "\uD83D\uDC28", "\uD83D\uDC12", "\uD83D\uDC35", "\uD83D\uDC06",
+            "\uD83D\uDC2f", "\uD83D\uDC3b", "\uD83D\uDC2a", "\uD83D\uDC2b", "\uD83D\uDC0a",
+            "\uD83D\uDC33", "\uD83D\uDC0b", "\uD83D\uDC1f", "\uD83D\uDC20", "\uD83D\uDC21",
+            "\uD83D\uDC19", "\uD83D\uDC1a", "\uD83D\uDC2c", "\uD83D\uDC0c", "\uD83D\uDC1b",
+            "\uD83D\uDC1c", "\uD83D\uDC1d", "\uD83D\uDC1e", "\uD83D\uDC32", "\uD83D\uDC09",
+            "\uD83D\uDC3e", "\uD83C\uDF78", "\uD83C\uDF7A", "\uD83C\uDF7b", "\uD83C\uDF77",
+            "\uD83C\uDF79", "\uD83C\uDF76", "\u2615", "\uD83C\uDF75", "\uD83C\uDF7c",
+            "\uD83C\uDF74", "\uD83C\uDF68", "\uD83C\uDF67", "\uD83C\uDF66", "\uD83C\uDF65",
+            "\uD83C\uDF70", "\uD83C\uDF6a", "\uD83C\uDF6b", "\uD83C\uDF6c", "\uD83C\uDF6d",
+            "\uD83C\uDF6e", "\uD83C\uDF6f", "\uD83C\uDF73", "\uD83C\uDF54", "\uD83C\uDF5f",
+            "\uD83C\uDF5d", "\uD83C\uDF55", "\uD83C\uDF56", "\uD83C\uDF57", "\uD83C\uDF64",
+            "\uD83C\uDF63", "\uD83C\uDF71", "\uD83C\uDF5e", "\uD83C\uDF5c", "\uD83C\uDF59",
+            "\uD83C\uDF5a", "\uD83C\uDF5b", "\uD83C\uDF72", "\uD83C\uDF65", "\uD83C\uDF62",
+            "\uD83C\uDF61", "\uD83C\uDF58", "\uD83C\uDF60", "\uD83C\uDF4c", "\uD83C\uDF4e",
+            "\uD83C\uDF4f", "\uD83C\uDF4a", "\uD83C\uDF4b", "\uD83C\uDF44", "\uD83C\uDF45",
+            "\uD83C\uDF46", "\uD83C\uDF47", "\uD83C\uDF48", "\uD83C\uDF49", "\uD83C\uDF50",
+            "\uD83C\uDF51", "\uD83C\uDF52", "\uD83C\uDF53", "\uD83C\uDF4d", "\uD83C\uDF30",
+            "\uD83C\uDF31", "\uD83C\uDF32", "\uD83C\uDF33", "\uD83C\uDF34", "\uD83C\uDF35",
+            "\uD83C\uDF37", "\uD83C\uDF38", "\uD83C\uDF39", "\uD83C\uDF40", "\uD83C\uDF41",
+            "\uD83C\uDF42", "\uD83C\uDF43", "\uD83C\uDF3a", "\uD83C\uDF3b", "\uD83C\uDF3c",
+            "\uD83C\uDF3d", "\uD83C\uDF3e", "\uD83C\uDF3f", "\u2600", "\uD83C\uDF08",
+            "\u26c5", "\u2601", "\uD83C\uDF01", "\uD83C\uDF02", "\u2614",
+            "\uD83D\uDCA7", "\u26a1", "\uD83C\uDF00", "\u2744", "\u26c4",
+            "\uD83C\uDF19", "\uD83C\uDF1e", "\uD83C\uDF1d", "\uD83C\uDF1a", "\uD83C\uDF1b",
+            "\uD83C\uDF1c", "\uD83C\uDF11", "\uD83C\uDF12", "\uD83C\uDF13", "\uD83C\uDF14",
+            "\uD83C\uDF15", "\uD83C\uDF16", "\uD83C\uDF17", "\uD83C\uDF18", "\uD83C\uDF91",
+            "\uD83C\uDF04", "\uD83C\uDF05", "\uD83C\uDF07", "\uD83C\uDF06", "\uD83C\uDF03",
+            "\uD83C\uDF0c", "\uD83C\uDF09", "\uD83C\uDF0a", "\uD83C\uDF0b", "\uD83C\uDF0e",
+            "\uD83C\uDF0f", "\uD83C\uDF0d", "\uD83C\uDF10"
+    };
+}
