+++ android-emojis/emoji-keyboard/src/main/res/layout/emoji_keyboard_layout.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<com.klinker.android.emoji_keyboard.view.EmojiKeyboardView
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/keyboard"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:layout_alignParentBottom="true"
+    android:keyPreviewLayout ="@layout/preview"
+    />
