+++ android-emojis/emoji-keyboard/src/main/res/layout/preview.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<TextView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:gravity="center"
+    android:background="#ffffff"
+    android:textStyle="bold"
+    android:textSize="30sp"
+    >
+</TextView>
